feat: move get_mut and get_meta function in messenger suffix to core suffix package
